addEvent(s) methods do the work, remove utils files
